Add files via upload
The Phonebook.py file contains the source code of the  Phonebook application in Python that includes the key modules: Insert contact, Search contact, Display contacts, Delete contact, Update contact, Sort contacts, and Analyze search efficiency. This can be run in VS Code.
The Phonebook_gui.py file contains the code of the GUI which was build using Tkinter, which is a popular library for building desktop applications in Python. This will allow users to interact with the app through buttons, text fields, and other UI elements instead of the command-line interface.