The Source Code
This is the source code for Phonebook application in Python that includes the key modules: Insert contact, Search contact, Display contacts, Delete contact, Update contact, Sort contacts, and Analyze search efficiency. This can be run in VS Code.
When you run the program, it will present the user  with a menu where they  can choose the operations they wish to perform.